Plate_recognizer/requirements.txt
opencv-python
scikit-image
scipy
fastapi[standard]
numpy
pillow
ultralytics==8.2.98
transformers==4.46.2
python-dotenv